Constructors

Index

HyGene.Context

HyGene.HyGeneModel

HyGene.LongTermMemory

HyGene.MemoryTrace

HyGene.ObsTrace

HyGene.SemanticMemory

HyGene.SetofContenders

Functions

HyGene.HyGeneModel — Type

HyGeneModel(context_labels::Vector{Symbol}, hypothesis_labels::Vector{Symbol},
            n_trace_vec::Vector{<:Number}, A_c::Number=0.2, t_max::Integer=10,
            n_features::Integer=15, focal_similarity::Float64=0.0, encoding_fidelity::Float64=0.75)
HyGeneModel(n_contexts::Integer, n_hypotheses::Integer, n_obs_per_proto::Integer,
            A_c::Number=0.2, t_max::Integer=10, n_features::Integer=15,
            focal_similarity::Float64=0.0, encoding_fidelity::Float64=0.75)

Constructor functions for the HyGeneModel object class.

HyGene.Context — Method

Context(label::Symbol, description::String, n_features::Integer)

Constructor for a new Context object. Generates a feature vector with values sampled from [-1.0, 0.0, 1.0].

Arguments

label::Symbol: Symbol indicating the overall type of Context. In simulations, similar Context structures will have the same label.

description::String: Optional description to help differentiate Context objects with similar labels.

n_features::Integer: Length of resulting feature vector.

Returns

Context object with randomly-generated feature vector of length n_features.

HyGene.MemoryTrace — Method

MemoryTrace(label::Symbol, description::String, info_vec::Vector{<:Information})

Constructor for a [MemoryTrace]@(def) object.

Arguments

label::Symbol: Symbol indicating the unique type of observation/trace.

description::String: Optional description to help differentiate traces with similar labels.

info_vec::Vector{<:Information}: List of Information/Context objects to be written to the resulting MemoryTrace object.

Returns

MemoryTrace object.

HyGene.ObsTrace — Method

ObsTrace(label::Symbol, description::String, info_vec::Vector{<:Information})

Constructor for a ObsTrace object.

Arguments

label::Symbol: Symbol indicating the unique type of observation/trace.

description::String: Optional description to help differentiate traces with similar labels.

info_vec::Vector{<:Information}: List of Information/Context objects to be written to the resulting ObsTrace object.

Returns

ObsTrace object.

HyGene.LongTermMemory — Method

LongTermMemory(trace_vec::Vector{<:Trace}, A_c::Float64)

Constructor for a LongTermMemory object.

Arguments

trace_vec::Vector{<:Trace}: List of traces to be written to long-term memory.

A_c::Float64: Activation threshold. Used during the echo_content and cond_echo_intensity operations.

Returns

LongTermMemory object.

HyGene.SemanticMemory — Method

SemanticMemory(ltm::LongTermMemory)
SemanticMemory(exemplars::Vector{<:Trace})

Constructor for the SemanticMemory structure class. If there are only contents in long-term memory, then this operation will create averaged feature vectors. If prototypes are given, then SemanticMemory will be populated with those.

Arguments

ltm::LongTermMemory: Memory store from which exemplar traces are to be extracted.

Returns

SemanticMemory object.

HyGene.SetofContenders — Method

SetofContenders(contenders::Vector{<:MemoryTrace}, t_max::Integer)
SetofContenders(t_max::Integer)

Constructor function for the SetofContenders object. SetofContenders objects are typically instantiated with no contents.

Arguments

contenders::Vector{<:MemoryTrace}: Vector of memory traces to populate the SoC.

t_max::Integer: Maximum number of retrieval failures.

Returns

SetofContenders object.