General Functions

Index

HyGene.bounce_memory!

HyGene.clear_content!

HyGene.cond_echo_intensity

HyGene.create_vec

HyGene.echo_content

HyGene.populate_soc!

HyGene.soc_posterior_prob

HyGene.standardize_trace!

HyGene.trace_activation

HyGene.trace_decay!

HyGene.trace_replication

HyGene.trace_similarity

Functions

HyGene.create_vec — Function

create_vec(n_values::Integer)

Returns a vector of length n_values with numbers randomly chosen from [-1.0 0.0 1.0].

Arguments

n_values::Integer: Number of features in resulting vector.

Example

julia> using Random; Random.seed!(1);

julia> create_vec(5)
5-element Vector{Float64}:
 -1.0
  0.0
  1.0
  0.0
  1.0

create_vec(n_vecs::Integer, n_values::Integer)

Returns an array of minivectors. The array is length n_vecs while each minivector has n_values features randomly sampled from [-1.0 0.0 1.0].

Arguments

n_vecs::Integer: Number of minivectors.

n_values::Integer: Number of features in each mini-vector.

Example

julia> using Random; Random.seed!(1);

julia> create_vec(2, 3)
2-element Vector{Vector{Float64}}:
 [0.0, -1.0, -1.0]
 [0.0, 0.0, -1.0]

create_vec(n_vecs::Integer, n_values::Integer, similarity::Number)

Returns an array of length n_vecs with each minivector containing n_values features randomly sampled from [-1.0 0.0 1.0]. After an initial minivector is created, the subsequent vectors will replicate each value with probability similarity.

Arguments

n_vecs::Integer: Number of minivectors.

n_values::Integer: Number of features in each mini-vector.

simiarlity::Number: Degree of similarity between each mini-vector.

Example

julia> using Random; Random.seed!(1);

julia> vecs = create_vec(4, 4, 0.5)
4-element Vector{Vector{Float64}}:
 [0.0, -1.0, -1.0, 1.0]
 [0.0, -1.0, 1.0, 1.0]
 [0.0, -1.0, 1.0, 1.0]
 [0.0, 1.0, 0.0, 1.0]

julia> vecs[1]'vecs[2] / length(vecs[1]) # dot product
0.25

HyGene.trace_replication — Function

trace_replication(trace::Vector{Number}, similarity::Number, decay::Bool=false)

Returns a degraded replication of a given trace vector. The elements of the returned vector are chosen from the original with a probabiliy of similarity. Values not copied from the original vector are randomly samped from [-1.0 0.0 1.0] or assigned 0.0 if you are decaying the trace (decay==true).

Arguments

trace::Vector{Number}: Vector of integers to be duplicated.

similarity::Number: Degree of similarity between the original trace and resulting vector.

decay::Bool: Indicates if the similarity argument should be treated as the degree of decay. If true, then similarity is inverted, i.e., subtraced from 1.

Example

julia> using Random; Random.seed!(1);

julia> vector = [1.0, 1.0, 0.0, -1.0];

julia> trace_replication(vector, 0.5)
4-element Vector{Float64}:
 1.0
 1.0
 0.0
 0.0

trace_replication(trace::Information, similarity::Number, decay::Bool=false)
trace_replication(trace::Trace, similarity::Number, decay::Bool=false)

Returns a degraded replication of a given object under the HypothesisGeneration` hierarchy. The elements of the minivectors within the returned object are chosen from the original minivectors with a probabiliy of similarity. Values not copied from the original vector are randomly samped from [-1.0 0.0 1.0]or assigned 0.0 if you are decaying the trace (decay==true)

The following HypothesisGeneration types can be used as the Trace object:

Trace
├─ Information
│  └─ Context
├─ MemoryTrace
└─ ObsTrace

Arguments

trace: Object of either Information or Trace to be replicated.

similarity::Number: Degree of similarity between the original trace and resulting object.

decay::Bool: Indicates if the similarity argument should be treated as the degree of decay. If true, then similarity is inverted, i.e., subtraced from 1.

Example

julia> using Random; Random.seed!(1);

julia> context = Context(:a, "this is context a", 5)
Context
  label: Symbol a
  description: String "this is context a"
  n_features: Integer 5
  contents: Array{Float64}((5,)) [-1.0, 0.0, 1.0, 0.0, 1.0]
  A_i: Float64 0.0

julia> trace_replication(context, 0.5)
Context
  label: Symbol a
  description: String "this is context a"
  n_features: Integer 5
  contents: Array{Float64}((5,)) [-1.0, 0.0, 1.0, 0.0, -1.0]
  A_i: Float64 0.0

HyGene.trace_activation — Function

trace_activation(probe::Vector, trace::Vector, exponent::Integer=3)
trace_activation(probe::HypothesisGeneration, trace::HypothesisGeneration, exponent::Integer=3)

The trace_similarity calculation accelerated by an exponent of degree exponent, usually 3. This is analogous to a cubed cosine similarity. See Eqs 1 and 2 in Thomas et al. (2008) [1].

\[\begin{equation} A_i = \left( \frac{\sum_{j=1}^n P_j T_{i,j}}{\sqrt{\sum_{j=1}^n P_j^2} \sqrt{\sum_{j=1}^n T_{i,j}^2}} \right) ^3 \end{equation}\]

See the trace_similarity documentation for the base operation.

Arguments

probe: An object to be compared to the trace.

trace: An object to be compared to the probe.

exponent::Integer=3: Degree of acceleration to the S_i value.

Example

julia> using Random; Random.seed!(25);

julia> context_a = Context(:a, "this is context a", 10);

julia> context_b = trace_replication(deepcopy(context_a), 0.5);

julia> trace_activation(context_a, context_b)
0.216

HyGene.trace_similarity — Function

trace_similarity(probe::Vector{<:Number}, trace::Vector{<:Number})

Calculates the similarity between a probe vector and a trace. Eq 1 in Thomas et al. (2008) [1]. Essentially a dot product.

\[\begin{equation} S_i = \frac{\sum_{j=1}^n P_j T_{i,j}}{\sqrt{\sum_{j=1}^n P_j^2} \sqrt{\sum_{j=1}^n T_{i,j}^2}} \end{equation}\]

Arguments

probe::Vector{<:Number}: A vector of integers to be compared to the trace.

trace::Vector{<:Number}: A vector of integers.

Example

julia> a = [1.0, -1.0, 0.0, 1.0];

julia> b = [1.0, 0.0, 0.0, 1.0];

julia> trace_similarity(a,b)
0.6666666666666666

trace_similarity(probe::Information, trace::Information)
trace_similarity(probe::Trace, trace::Trace)

Calculates the similarity between a probe HypothesisGeneration object and a trace object. If there are multiple Context vectors in the probe and trace, it will recursively compute trace_similarity for each vector and apply it to the overall S_i value.

\[\begin{equation} S_i = \frac{\sum_{j=1}^n P_j T_{i,j}}{\sqrt{\sum_{j=1}^n P_j^2} \sqrt{\sum_{j=1}^n T_{i,j}^2}} \end{equation}\]

The following types are accepted as probe and trace objects:

Trace
├─ Information
│  └─ Context
├─ MemoryTrace
└─ ObsTrace

Arguments

probe: An object of either Information or Trace to be compared to trace.

trace: An object of either Information or Trace.

Example

julia> using Random; Random.seed!(25);

julia> context_a = Context(:a, "this is context a", 10);

julia> context_b = trace_replication(context_a, 0.5);

julia> trace_similarity(context_a, context_b)
0.6

HyGene.trace_decay! — Function

trace_decay!(trace::Information, decay::Number)
trace_decay!(trace::Trace, decay::Number)
trace_decay!(store::MemoryStore, decay::Number)

Returns a HypothesisGeneration object with constituent Information vectors of numbers that have been copied to the new object with probability 1 - decay.

Arguments

trace: An Information, Trace, or MemoryStore object with contents to be degraded.

decay::Number: The degree of decay to the trace elements.

Example

julia> using Random; Random.seed!(2);

julia> context_a = Context(:a, "this is context a", 10);

julia> context_b = deepcopy(context_a);

julia> trace_decay!(context_b, 0.5);

julia> trace_similarity(context_a, context_b)
0.6

HyGene.bounce_memory! — Function

bounce_memory!(observation::Trace, memory::MemoryStore)

Runs the trace_activation command between the observation object and all objects in the memory object. Activation values A_i and then written back to the Information vectors in memory.

This operation is used in parallel with echo_content and SetofContenders functions.

Arguments

observation::Trace: Observation from the real world to be compared to memory.

memory::MemoryStore: Memory store that the observation is compared to.

HyGene.clear_content! — Function

clear_content!(context::Information)
clear_content!(trace::Trace)

Resets Information vectors and activation (A_i) values back to zeros. No other fields, such as description are affected. Mostly useful for running multiple simulations.

Arguments

An Information or Trace object to be reset.

Example

julia> using Random; Random.seed!(10);

julia> context_a = Context(:a, "this is context a", 4)
Context
  label: Symbol a
  description: String "this is context a"
  n_features: Int64 4
  contents: Array{Float64}((4,)) [-1.0, -1.0, -1.0, -1.0]
  A_i: Float64 0.0

julia> clear_content!(context_a);

julia> context_a
Context
  label: Symbol a
  description: String "this is context a"
  n_features: Int64 4
  contents: Array{Float64}((4,)) [0.0, 0.0, 0.0, 0.0]
  A_i: Float64 0.0

HyGene.standardize_trace! — Function

standardize_trace!(trace::Trace)

Divides elements of all Information objects by the largest abolute value across all objects, restricting values in the range of [-1.0, 1.0]. Helps resolve the ambiguous recall problem in global memory models (Hintzman, 1986) [2].

Arguments

trace: a Trace object with unstandardized Context mini-vectors.

julia> using Random; Random.seed!(5);

julia> context_a = Context(:a, "this is context a", 5);

julia> context_a.contents .*= rand(5)
5-element Vector{Float64}:
  0.4113527070924652
  0.8090056332338996
 -0.4913697257138979
  0.0
  0.0

julia> trace_a = MemoryTrace(:event_a, "this is event a", [context_a]);

julia> standardize_trace!(trace_a);

julia> trace_a.contents[1].contents
5-element Vector{Float64}:
  0.5084670491701446
  1.0
 -0.6073749125203349
  0.0
  0.0

HyGene.echo_content — Function

echo_content(observation::Trace, memory::MemoryStore, 
             conditional::Bool=true, standardize::Bool=true)

Computes trace_activation between the observation object and all objects in the memory store. Information vectors that are above activation threshold memory.A_c are then multiplied against their conditional activation values A_i (if conditional==true) and then summed across columns. The result is the "echo content" from a given memory store, i.e., a MemoryTrace object with float values in the Information objects. The standardize parameter controls whether the operation controls for Hintzman's (1986) [2] ambiguous recall problem.

The resulting conditional echo content is given by Eq 7 in Thomas et al. (2008) [1]:

\[\begin{equation} C_c = \sum_{i=1}^K A_i T_{i,j} \end{equation}\]

Arguments

observation::Trace: Observation made in the real world, usually a copy of a prototype.

memory::MemoryStore: A MemoryStore object by which the observation is "bounced" off of.

conditional::Bool=true: Flag that indicates whether or not the program should calculate conditional echo content. true by default.

standardized::Bool=true: Flat that indicates whether or not the echo content should be standardized.

HyGene.populate_soc! — Function

populate_soc!(echo::MemoryTrace, soc::SetofContenders, semantic::SemanticMemory)
populate_soc!(echo::MemoryTrace, model::HyGeneModel)

This operation compares a trace (echo) to contents in semantic memory (semantic) and writes semantic traces with activation values above a dynamic threshold (soc.act_min) to the SetofContenders object.

Semantic traces are randomly sampled, and those that are above threshold will get written to the SetofContenders and the activation threshold (soc.act_min) will be set to that trace's activation (trace.A_i). If a trace's activation value does not exceed the threshold, then it will be considered a retrieval failure and will increment the failure count variable (t) by 1.

The function will continue to randomly sample traces with replacement until the maximum number of retrieval attempts (t_max) has been reached.

Arguments

echo::MemoryTrace: Echo content trace to be compared to SemanticMemory.

soc::SetofContenders: Working memory store holding the contenders.

semantic::SemanticMemory: SemanticMemory store with trace exemplars.

model::HyGeneModel: A generic Model object with a SetofContenders and SemanticMemory store.

HyGene.cond_echo_intensity — Function

cond_echo_intensity(probe::MemoryTrace, memory::MemoryStore, conditional::Bool=true)

Returns the conditional echo intensity of a memory trace object (probe) against a memory store (memory). Echo intensity values are computed against the contents of the probe for each item in memory. All intensity values above a threshold A_c are then summed and divided by the number of above-threshold activations to produce I_c. If conditional is set to false, the function will return the sum of all intensity values divided by the number of items with activations above zero.

As per Eq. 5 in Thomas et al. (2008) [1]:

\[\begin{equation} I_c = \frac{\sum I_{A_i \geq A_c}}{K} \end{equation}\]

Arguments

probe: A Trace object to be compared to the memory store.

memory: A MemoryStore object with traces used to generate echo intensity.

conditional::Bool=true: Flag to indicate whether the function should return conditional echo intensity. If false, then it will only return echo intensity and not conditional echo intensity.

HyGene.soc_posterior_prob — Function

soc_posterior_prob(soc, memory)
soc_posterior_prob(model)

Returns a vector of posterior probabilities for all leading contenders in the SetofContenders. This is accomplished by computing conditional echo intensity values for all contenders against LongTermMemory and dividing them by the sum.

As per Eq. 8 in Thomas et al. (2008) [1]:

\[\begin{equation} P(H_i | D_{obs}) = \frac{I_{C_i}}{\sum_{i=1}^w I_{C_i}} \end{equation}\]

1Thomas, R. P., Dougherty, M. R., Sprenger, A. M., & Harbison, J. (2008). Diagnostic hypothesis generation and human judgment. Psychological Review, 115(1), 155-185. https://psycnet.apa.org/doi/10.1037/0033-295X.115.1.155

2Hintzman, D. L. (1986). "Schema abstraction" in a multiple-trace memory model. Psychological Review, 93(4), 411-428. https://psycnet.apa.org/doi/10.1037/0033-295X.93.4.411