Simulation Controllers

Index

HyGene.create_labels

HyGene.create_observation

HyGene.create_prototypes

HyGene.create_traces

Functions

HyGene.create_labels — Function

create_labels(n_labels::Integer, label::String)

Creates generic labels. For use with create_prototypes.

Arguments

n_labels::Integer: The desired number of labels to be generated.

label::String: A generic label to prepend other identifying data.

Example

julia> create_labels(2, "example")
2-element Vector{Symbol}:
 :example1
 :example2

HyGene.create_prototypes — Function

create_prototypes(context_labels::Vector{Symbol}, hypothesis_labels::Vector{Symbol},
                  n_features::Integer=15, focal_sim::Number=0.0)

Creates a series of prototypes. First, a focal prototype is produced. Next, alternative hypothesis prototypes are generated with the features copied from the focal prototype with probability focal_sim. Labels are added to the prototypes and constituent Context objects.

Arguments

context_labels::Vector{Symbol}: A list of labels (of type Symbol) for the generated contexts within prototypes, i.e., one for each context.

hypothesis_labels::Vector{Symbol}: A list of labels (of type Symbol) for the generated prototypes/hypotheses, i.e., one for each object.

n_features::Integer: Number of features in each Context minivector.

focal_sim::Number=0.0: Degree of similarity between the focal hypothesis and alternative hypotheses. Set to no (0.0) similarity by default.

create_prototypes(n_contexts::Integer, n_hypotheses::Integer, n_features::Integer=15, 
                focal_sim::Number=0.0)

Prototype creation script in the case where labels are not given by the user. In this case, labels are automatically generated by the program and added to the hypothesis/prototype objects.

Arguments

n_contexts::Integer: The number of Context objects within each hypothesis object.

n_hypotheses::Integer: The number of hypotheses to be generated.

n_features::Integer: Number of features in each Context minivector.

focal_sim::Number=0.0: Degree of similarity between the focal hypothesis and alternative hypotheses. Set to no (0.0) similarity by default.

HyGene.create_traces — Function

create_traces(prototypes::Vector{MemoryTrace}, n_trace_vec::Vector{<:Number},
              encode_fidelity::Number=0.75)

Generates MemoryTrace objects to be written to LongTermMemory. Each trace is copied from a prototype with similarity equal to encode_fidelity. The number of copies of each prototype is set by a vector of integers (n_trace_vec).

Arguments

prototypes::Vector{MemoryTrace}: List of prototypes to be copied into LongTermMemory.

n_trace_vec::Vector{<:Number}: A vector of numbers indicating how many times each prototype should be copied. For example, 3 prototypes could be copied with frequencies [2,4,2] each.

encode_fidelity::Number=0.75: The degree to which traces should match their respective prototypes.

create_traces(prototypes::Vector{MemoryTrace}, n_traces::Integer,
              encode_fidelity::Number=0.75)

Generates MemoryTrace objects to be written to LongTermMemory. Each trace is copied from a prototype with similarity equal to encode_fidelity. Each prototype is copied n_traces times.

Arguments

prototypes::Vector{MemoryTrace}: List of prototypes to be copied into LongTermMemory.

n_traces::Number: The number of copies of each prototype.

encode_fidelity::Number=0.75: The degree to which traces should match their respective prototypes.

HyGene.create_observation — Function

create_observation(prototype::MemoryTrace, d_obs_sim::Number=0.85)

Generates an observation from the real world, i.e., ObsTrace. Each observation is copied from a prototype with similarity equal to d_obs_sim.

Arguments

prototype::MemoryTrace: Prototype to be copied.

d_obs_sim::Number=0.85: Degree to which the observation matches the prototype.