Structures

Index

HyGene.Context

HyGene.HyGeneModel

HyGene.HypothesisGeneration

HyGene.Information

HyGene.LongTermMemory

HyGene.MemoryStore

HyGene.MemoryTrace

HyGene.Model

HyGene.ObsTrace

HyGene.SemanticMemory

HyGene.SetofContenders

HyGene.Trace

Abstract Types

HyGene.HypothesisGeneration — Type

HypothesisGeneration <: ContinuousUnivariateDistribution

An abstract type representing all structures related to HyGene simulations. Mostly useful when using custom samplers, such as with Turing.jl, but also for preventing excessive function re-definition for different sub-types/structures.

Type Hierarchy

HypothesisGeneration
├─ MemoryStore
│  ├─ LongTermMemory
│  ├─ SemanticMemory
│  └─ SetofContenders
├─ Model
│  └─ HyGeneModel
└─ Trace
   ├─ Information
   │  └─ Context
   ├─ MemoryTrace
   └─ ObsTrace

HyGene.Model — Type

Model <: HypothesisGeneration

An abstract type representing large simulation objects. The base code only includes one Model type (HyGeneModel), but users can create new objects for bespoke simulations.

Subtypes

HyGeneModel

HyGene.MemoryStore — Type

MemoryStore <: HypothesisGeneration

An abstract type representing all memory stores.

Subtypes

LongTermMemory

SemanticMemory

SetofContenders

HyGene.Trace — Type

Trace <: MemoryStore

An abstract type representing chunks of information held in MemoryStore objects.

Subtypes

MemoryTrace

ObsTrace

HyGene.Information — Type

Information <: Trace

An abstract type representing information held within chuncks in MemoryStore objects. There is only one subtype in this abstract class, but more can be created for bespoke simulations.

Subtypes

Context

Model Objects

HyGene.HyGeneModel — Type

HyGeneModel <: Model

Object holding HypothesisGeneration simulation parameters, variables, and outcomes. The top-level operations that govern the HyGeneModel object are located in sim_controllers.jl.

Parameters

A_c::Float64: Activation threshold in LongTermMemory.

t_max::Integer: Maximum number of retrieval failures during SetofContenders operations.

decay::Float64: Degree of trace decay for LongTermMemory contents.

focal_similarity::Float64: Degree of similarity between the focal and alternative hypotheses.

encoding_fidelity::Float64: Degree of change between an observation in the environment and a new MemoryTrace object.

n_features::Integer: Length of context mini-vectors. 10 is recommended.

n_unique_contexts::Integer: Number of unique Context types.

n_unique_events::Integer: Number of unique MemoryTrace types in LongTermMemory and length of SemanticMemory store contents.

n_obs_per_event::Any: Number of unqiue observations per unique MemoryTrace types in LongTermMemory.

contexts::Vector{Symbol}: List of labels of unique types of Context objects.

hypotheses::Vector{Symbol}: List of labels of unique types of MemoryTrace objects.

len_ltm::Integer: Number of items in the LongTermMemory store. Should be n_unique_events * n_obs_per_event.

len_sm::Integer: Number of items in the SemanticMemory store. Should be n_unique_events.

prototypes::Vector{<:Trace}: List of hypothesis exemplars.

long_term_memory::LongTermMemory: A single LongTermMemory object.

semantic_memory::SemanticMemory: A single SemanticMemory object.

working_memory::SetofContenders: A single SetofContenders object.

Mutable Structures

HyGene.Context — Type

Context <: Information

Container used to hold information about a single "context", generally defined. Context structures are mutable and hold one numeric vector. They are considered to be the lowest level of information in the HypothesisGeneration environment.

Parameters

label::Symbol: Symbol indicating the overall type of Context. In simulations, similar Context structures will have the same label.

description::String: Optional description to help differentiate Context objects with similar labels.

n_features::Integer: Length of feature vector.

contents::AbstractArray: A single array of length n_features.

A_i::Float64: Optional value reflecting activation of the context. Typically unused during simulations.

Constructor

Context(label::Symbol, description::String, n_features::Integer)

HyGene.MemoryTrace — Type

MemoryTrace <: Trace

Container for objects that populate MemoryStore objects. They consist of 1 or more Context objects. These structures are analogous to information and memories that can potentially degrade across time. LongTermMemory stores typically contain multiple permutations of specific MemoryTrace objects while SemanticMemory holds single MemoryTrace objects for each type of observation. All parameters of the MemoryTrace object are mutable.

Parameters

label::Symbol: Symbol indicating the unique type of observation/trace.

description::String: Optional description to help differentiate traces with similar labels.

n_contexts::Integer: Number of unique types of Context objects. This is determined by finding unique label values across all Context objects.

contexts::Vector{Symbol}: List of unique types of Context objects.

contents::Vector{<:Information}: Contents of the memory store, i.e., 1 or more Context objects.

A_i::Float64: Activation value. Intitially set to 0.0 and then updated using the bounce_memory! function.

Constructor

MemoryTrace(label::Symbol, description::String, info_vec::Vector{<:Information})

HyGene.ObsTrace — Type

ObsTrace <: Trace

Structurally similar to the MemoryTrace object type, but represents an observation made in the "real world" that is then compared to a MemoryStore object. Typically unneeded, but can be useful for operations in parametric functions that differentiate memory traces and traces derived from outward observations.

Parameters

label::Symbol: Symbol indicating the unique type of observation/trace.

description::String: Optional description to help differentiate traces with similar labels.

n_contexts::Integer: Number of unique types of Context objects. This is determined by finding unique label values across all Context objects.

contexts::Vector{Symbol}: List of unique types of Context objects.

contents::Vector{<:Information}: Contents of the memory store, i.e., 1 or more Context objects.

A_i::Float64: Activation value. Typically unused.

Constructor

ObsTrace(label::Symbol, description::String, info_vec::Vector{<:Information})

HyGene.LongTermMemory — Type

LongTermMemory <: MemoryStore

A long-term memory store for Trace objects. Contents in LongTermMemory are potentially subject to degradation (trace_decay!). LongTermMemory is key in echo content calculations and posterior probability estimates resulting from the SoC. Items in this store are used to create exemplars in SemanticMemory.

Parameters

n_items::Integer: The number of Trace items in storage.

A_c::Float64: Activation threshold for conditional retrieval/probability. Used during the echo_content and cond_echo_intensity operations.

unique_contexts::Vector{Symbol}: List of unique symbols extracted from Trace objects.

contents::Vector{<:Trace}: List of traces in long-term memory.

Constructor

LongTermMemory(trace_vec::Vector{<:Trace}, A_c::Float64)

HyGene.SemanticMemory — Type

SemanticMemory <: MemoryStore

A memory store consisting of Trace objects. All of the items in SemanticMemory are exemplars derived from multiple observations of individual events in LongTermMemory. For example, in a LongTermMemory store with 10 separate traces with label :event_a and 10 separate traces with label :event_b, SemanticMemorywould be populated with two separate traces for:eventaand:eventb` with their item vectors being the means of the contents of the 10 separate traces for each event.

Parameters

n_items::Integer: The number of unique exemplars, or semantic memory traces.

content_labels::Vector{Symbol}: List of the labels for the unique exemplars.

contents::Vector{<:Trace}: List of unique exemplars in semantic memory.

A_c::Float64: Activation threshold for condition retrieval/probability. Not typically used in these simulations.

Constructor

SemanticMemory(ltm::LongTermMemory)

HyGene.SetofContenders — Type

SetofContenders <: MemoryStore

An analog to the "Working Memory" concept in Thomas et al.'s (2008) [1] original HyGene paper. The SetofContenders object is used to hold above-threshold traces from SemanticMemory when an outward observation (ObsTrace) is compared against its contents. The contents of the SetofContenders is then used to compute the probability that a given semantic representation is the correct exemplar, given the probe (cond_echo_intensity).

Parameters

n_contenders::Integer:

contenders::Vector{<:MemoryTrace}:

act_min::Float64:

t_max::Integer:

t::Integer:

Constructor

SetofContenders(t_max::Integer)